small refactor of image upload code to separate it for easier testing/stubbing
